implemented optional DCI event "all thresholds rearmed"; consistent use of int instead of long for event code on Java side

+++ src/client/nxmc/java/src/main/java/org/netxms/nxmc/modules/events/views/EventProcessingPolicyEditor.java
@@ -1095,7 +1095,7 @@ public class EventProcessingPolicyEditor extends ConfigurationView
          return true;
 
       // check event names
-      for(Long code : rule.getEvents())
+      for(Integer code : rule.getEvents())
       {
          EventTemplate evt = session.findEventTemplateByCode(code);
          if ((evt != null) && evt.getName().toLowerCase().contains(filterText))
